Refactor : 프로젝트 페이지 관련 컴포넌트 폴더 구조 개편
