+++ src/components/projects/navigation/ProjectPageNavigation.tsx
@@ -9,9 +9,9 @@ import PersonSearchIcon from '@mui/icons-material/PersonSearch';
 import VolunteerActivismIcon from '@mui/icons-material/VolunteerActivism';
 import HowToVoteIcon from '@mui/icons-material/HowToVote';
 import GroupsIcon from '@mui/icons-material/Groups';
-import useUserId from '../../hooks/reduxHooks/useUserId';
+import useUserId from '../../../hooks/reduxHooks/useUserId';
 
-const ProjectNavigation: React.FC = () => {
+const ProjectPageNavigation: React.FC = () => {
   const location = useLocation();
   const userId: number | null = useUserId();
 
@@ -137,4 +137,4 @@ const navLinkIcon = css`
   }
 `;
 
-export default ProjectNavigation;
+export default ProjectPageNavigation;
