+++ src/components/projects/recruitment/EmptyBox.tsx
