+++ src/pages/users/UserInfoPage.tsx
@@ -5,7 +5,7 @@ import { useMediaQuery } from 'react-responsive';
 import useToken from '../../hooks/reduxHooks/useToken';
 
 import Header from '../../components/common/Header';
-import ProjectNavigation from '../../components/projects/ProjectNavigation';
+import ProjectPageNavigation from '../../components/projects/navigation/ProjectPageNavigation';
 import UserInfoSection from '../../components/user/UserInfoSection';
 import Footer from '../../components/common/Footer';
 
@@ -22,7 +22,7 @@ const UserInfoPage: React.FC = () => {
   return (
     <>
       <Header />
-      <ProjectNavigation />
+      <ProjectPageNavigation />
       <UserInfoSection />
       {!smallMedia && <Footer />}
     </>
